Fix int16 overflow in BT.601 YUV->RGB conversion
The Y/U/V planes were cast to int16, but the BT.601 fixed-point sums
multiply Y by 256. For any Y >= 128 that exceeds int16's positive
range (256*128 = 32768 > 32767), wraps to a large negative value, and
the subsequent >> 8 leaves a substantial negative bias in the R, G, B
sums. The clip-to-[0,255] at the end then renders bright regions as
dark blue / black / magenta depending on which channel had enough
positive UV contribution to survive.

This matches the symptoms exactly: deep saturated sky with no detail
(Y high, R+G clipped 0, B partially recovered by 454*(U-128)), random
magenta speckle in mid-Y regions where overflow flips channels
independently, and clean rendering in low-Y areas like the dashboard
(where 256*Y stays under 32767).

Cast Y, U, V to int32 instead. Credit: noticed by Nizar.

+++ sunnypilot/nkaoud_nav/adjacent_vehicle_detector.py
@@ -276,8 +276,12 @@ class VisualVehicleDetector:
 
       # Slice buf.data as a memoryview FIRST, then wrap with np.array. This is
       # what snapshot.py does and it sidesteps any shape buf.data might carry.
+      # int32 (not int16) is required: subsequent 256*Y overflows int16 for any
+      # Y >= 128 (256*128 = 32768 > int16 max 32767), wraps negative, and after
+      # >> 8 ends up as a large negative bias in the BT.601 sums -- which is
+      # exactly what was making bright regions render as dark / magenta.
       y = np.array(buf.data[:uv_offset], dtype=np.uint8) \
-            .reshape((-1, stride))[:height, :width].astype(np.int16)
+            .reshape((-1, stride))[:height, :width].astype(np.int32)
       uv_data = buf.data[uv_offset:uv_offset + uv_plane_size]
       # The Spectra/BPS chroma output is actually NV21 (V,U interleaved) on
       # the wide road camera even though spectra.cc configures CAM_FORMAT_NV12
@@ -286,9 +290,9 @@ class VisualVehicleDetector:
       # snapshot.py has the same bug but happens to be used only on near-gray
       # dashboards where U~=V~=128 so the swap is invisible.
       v = np.array(uv_data[::2], dtype=np.uint8) \
-            .reshape((-1, stride // 2))[:height // 2, :width // 2].astype(np.int16)
+            .reshape((-1, stride // 2))[:height // 2, :width // 2].astype(np.int32)
       u = np.array(uv_data[1::2], dtype=np.uint8) \
-            .reshape((-1, stride // 2))[:height // 2, :width // 2].astype(np.int16)
+            .reshape((-1, stride // 2))[:height // 2, :width // 2].astype(np.int32)
       return y, u, v
     except Exception:
       cloudlog.exception("visual vehicle detector failed frame plane extraction")
